chore: update seeding for new signature
